requirements.txt
opencv_contrib_python==4.12.0.88
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==11.3.0
